refactor(sdk): extract QuotedTransferProvider abstraction from WarpCore
The EVM-specific offchain-quoting branch in `WarpCore.getTransferRemoteTxs`
moved behind a protocol-agnostic `QuotedTransferProvider` interface;
`EvmQuotedTransferProvider` now owns what was inlined in
`getQuotedCallsTransferTxs` + `resolveQuotedCallsParams` +
`getQuotedTransferFee`. WarpCore dispatches with one branch; the legacy
`quotedCalls?` param stays as backcompat sugar. No behavior change — the
572 existing sdk unit tests (including QuotedCalls coverage) pass
unchanged. Sets up the dispatch hook for the SVM offchain-quoting provider.
