+++ typescript/sdk/src/warp/WarpCore.ts
@@ -5,7 +5,6 @@ import {
   HexString,
   Numberish,
   ProtocolType,
-  addressToBytes32,
   assert,
   convertDecimalsToIntegerString,
   convertToProtocolAddress,
@@ -15,6 +14,7 @@ import {
   rootLogger,
 } from '@hyperlane-xyz/utils';
 import { Keypair } from '@solana/web3.js';
+import { isAddress } from 'viem';
 
 import type { PredicateAttestation } from '../predicate/PredicateApiClient.js';
 import type { MultiProviderAdapter } from '../providers/MultiProviderAdapter.js';
@@ -47,17 +47,11 @@ import {
   isHypCrossCollateralAdapter,
   isPredicateCapableAdapter,
 } from '../token/adapters/ITokenAdapter.js';
-import {
-  buildExecuteCalldata,
-  buildQuoteCalldata,
-} from '../quoted-calls/builder.js';
+import { EvmQuotedTransferProvider } from '../quoted-calls/EvmQuotedTransferProvider.js';
+import type { QuotedTransferProvider } from '../quoted-calls/QuotedTransferProvider.js';
 import type { Quote } from '../quoted-calls/codec.js';
-import {
-  decodeQuoteExecuteResult,
-  extractQuoteTotals,
-} from '../quoted-calls/codec.js';
 import type { QuotedCallsParams } from '../quoted-calls/types.js';
-import { TokenPullMode } from '../quoted-calls/types.js';
+import { resolveDestinationToken } from './resolveDestinationToken.js';
 import { ChainName, ChainNameOrId } from '../types.js';
 import { messageAmountFromLocal } from '../utils/decimals.js';
 
@@ -477,6 +471,7 @@ export class WarpCore {
     attestation,
     destinationToken,
     quotedCalls,
+    quotedTransfer,
   }: {
     originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
@@ -487,28 +482,42 @@ export class WarpCore {
     /** Optional Predicate attestation for compliance-gated warp routes */
     attestation?: PredicateAttestation;
     destinationToken?: IToken;
-    /** When provided, builds an atomic QuotedCalls.execute() tx instead of separate approve+transfer */
+    /**
+     * Backcompat sugar — wraps into an `EvmQuotedTransferProvider` internally.
+     * New callers should prefer `quotedTransfer` (protocol-agnostic).
+     */
     quotedCalls?: QuotedCallsParams;
+    /**
+     * Protocol-agnostic offchain-quoting hook. EVM and Sealevel provide their
+     * own implementations; WarpCore dispatches via a single branch.
+     */
+    quotedTransfer?: QuotedTransferProvider;
   }): Promise<Array<WarpTypedTransaction>> {
-    // QuotedCalls and attestation are mutually exclusive: the QuotedCalls.execute() path
-    // calls transferRemote, not transferRemoteWithAttestation. Composing them would require
-    // new contract support; for now, surface a clear error rather than silently dropping
-    // the attestation.
+    const provider =
+      quotedTransfer ??
+      (quotedCalls ? new EvmQuotedTransferProvider(quotedCalls) : undefined);
+
+    // Quoted transfers and attestation are mutually exclusive: the quoted path
+    // calls transferRemote, not transferRemoteWithAttestation. Composing them
+    // would require new contract support; surface a clear error rather than
+    // silently dropping the attestation.
     assert(
-      !(quotedCalls && attestation),
-      'quotedCalls and attestation cannot be used together. The QuotedCalls path does not support attestation-gated transfers.',
+      !(provider && attestation),
+      'quoted transfer and attestation cannot be used together. The quoted path does not support attestation-gated transfers.',
     );
 
-    // QuotedCalls atomic path — single execute() tx with quotes + token pull + transfer + sweep
-    if (quotedCalls) {
-      return this.getQuotedCallsTransferTxs({
+    if (provider) {
+      assert(
+        provider.protocol === originTokenAmount.token.protocol,
+        `Quoted-transfer provider protocol (${provider.protocol}) does not match origin token protocol (${originTokenAmount.token.protocol})`,
+      );
+      return provider.buildQuotedTransferTxs({
+        warpCore: this,
         originTokenAmount,
         destination,
         sender,
         recipient,
-        quotedCalls,
         destinationToken,
-        feeQuotes: quotedCalls.feeQuotes,
       });
     }
 
@@ -855,84 +864,11 @@ export class WarpCore {
   }
 
   /**
-   * Resolve common params for QuotedCalls operations.
-   */
-  protected resolveQuotedCallsParams({
-    originTokenAmount,
-    destination,
-    recipient,
-    quotedCalls,
-    destinationToken,
-  }: {
-    originTokenAmount: TokenAmount<IToken>;
-    destination: ChainNameOrId;
-    recipient: Address;
-    quotedCalls: QuotedCallsParams;
-    destinationToken?: IToken;
-  }) {
-    const { token, amount } = originTokenAmount;
-    assert(
-      isEVMLike(token.protocol),
-      'QuotedCalls is only supported on EVM origins',
-    );
-    assert(!token.isNft(), 'QuotedCalls does not support NFT routes');
-
-    const destinationDomainId = this.multiProvider.getDomainId(destination);
-    // For collateral routes, the ERC20 token is collateralAddressOrDenom.
-    // For synthetic/native routes, use addressOrDenom (the router itself).
-    // Only treat as native (zeroAddress) if collateral is explicitly address(0).
-    const collateral = token.collateralAddressOrDenom;
-    const tokenAddress = (
-      collateral && !isZeroishAddress(collateral)
-        ? collateral
-        : token.isNative() || token.isHypNative()
-          ? '0x0000000000000000000000000000000000000000'
-          : token.addressOrDenom
-    ) as `0x${string}`;
-
-    let targetRouter: `0x${string}` | undefined;
-    if (
-      destinationToken &&
-      this.isCrossCollateralTransfer(token, destinationToken)
-    ) {
-      const resolved = this.resolveDestinationToken({
-        originToken: token,
-        destination,
-        destinationToken,
-      });
-      assert(
-        resolved.addressOrDenom,
-        'Destination token missing addressOrDenom for cross-collateral',
-      );
-      targetRouter = addressToBytes32(resolved.addressOrDenom) as `0x${string}`;
-    }
-
-    return {
-      quotedCallsAddress: quotedCalls.address,
-      warpRoute: token.addressOrDenom as `0x${string}`,
-      destination: destinationDomainId,
-      recipient: addressToBytes32(recipient) as `0x${string}`,
-      amount,
-      token: tokenAddress,
-      quotes: quotedCalls.quotes,
-      clientSalt: quotedCalls.clientSalt,
-      targetRouter,
-    };
-  }
-
-  /**
    * Quote fees for a QuotedCalls transfer via quoteExecute eth_call.
-   * Returns structured fee data (like getInterchainTransferFee) plus
-   * the raw Quote[][] needed to build the execute tx.
+   * Thin delegate to `EvmQuotedTransferProvider.getQuotedTransferFee` —
+   * kept on WarpCore so existing callers (CLI, UI) stay working.
    */
-  async getQuotedTransferFee({
-    originTokenAmount,
-    destination,
-    sender,
-    recipient,
-    quotedCalls,
-    destinationToken,
-  }: {
+  async getQuotedTransferFee(args: {
     originTokenAmount: TokenAmount<IToken>;
     destination: ChainNameOrId;
     sender: Address;
@@ -942,181 +878,25 @@ export class WarpCore {
   }): Promise<{
     igpQuote: TokenAmount<IToken>;
     tokenFeeQuote?: TokenAmount<IToken>;
-    /** Raw per-command quotes — pass to getTransferRemoteTxs */
     feeQuotes: Quote[][];
   }> {
-    const { token: originToken } = originTokenAmount;
-    const originName = originToken.chainName;
-
-    const transferParams = this.resolveQuotedCallsParams({
-      originTokenAmount,
-      destination,
-      recipient,
-      quotedCalls,
-      destinationToken,
-    });
-
-    const quoteTx = buildQuoteCalldata(transferParams);
-    const provider = this.multiProvider.getEthersV5Provider(originName);
-    const quoteResult = await provider.call({
-      to: quoteTx.to,
-      data: quoteTx.data,
-      from: sender,
-    });
-    const feeQuotes = decodeQuoteExecuteResult(quoteResult as `0x${string}`);
-    const { nativeValue, tokenTotals } = extractQuoteTotals(feeQuotes);
-
-    // Build structured fee amounts matching getInterchainTransferFee return shape.
-    // For native routes, quoteTransferRemote includes the transfer amount in
-    // the native quotes, so we subtract it to get the fee-only portion.
-    const isNativeRoute = isZeroishAddress(transferParams.token);
-    const nativeToken = Token.FromChainMetadataNativeToken(
-      this.multiProvider.getChainMetadata(originName),
-    );
-    const igpFeeOnly = isNativeRoute
-      ? nativeValue - originTokenAmount.amount
-      : nativeValue;
-    const igpQuote = new TokenAmount(igpFeeOnly, nativeToken);
-
-    // Token fees = total ERC20 quoted minus the transfer amount
-    // sumQuotesByToken normalizes keys to lowercase
-    const tokenKey = transferParams.token.toLowerCase() as `0x${string}`;
     assert(
-      tokenTotals.size <= 1,
-      `Unexpected multi-token fee quotes: ${[...tokenTotals.keys()].join(', ')}`,
+      isAddress(args.sender),
+      `QuotedCalls sender must be a valid EVM address: ${args.sender}`,
     );
-    let tokenFeeQuote: TokenAmount<IToken> | undefined;
-    const totalTokenQuoted = tokenTotals.get(tokenKey);
-    if (totalTokenQuoted != null) {
-      const feeOnly = totalTokenQuoted - originTokenAmount.amount;
-      assert(
-        feeOnly >= 0n,
-        `Token fee quote underflow: quoted ${totalTokenQuoted} < amount ${originTokenAmount.amount}`,
-      );
-      if (feeOnly > 0n) {
-        tokenFeeQuote = new TokenAmount(feeOnly, originToken);
-      }
-    }
-
-    return { igpQuote, tokenFeeQuote, feeQuotes };
-  }
-
-  /**
-   * Build transactions for a QuotedCalls atomic transfer.
-   * Returns [approval (if needed), execute] transactions.
-   *
-   * @param feeQuotes Raw Quote[][] from getQuotedTransferFee.
-   *   If not provided, calls quoteExecute internally.
-   */
-  protected async getQuotedCallsTransferTxs({
-    originTokenAmount,
-    destination,
-    sender,
-    recipient,
-    quotedCalls,
-    destinationToken,
-    feeQuotes,
-  }: {
-    originTokenAmount: TokenAmount<IToken>;
-    destination: ChainNameOrId;
-    sender: Address;
-    recipient: Address;
-    quotedCalls: QuotedCallsParams;
-    destinationToken?: IToken;
-    feeQuotes?: Quote[][];
-  }): Promise<Array<WarpTypedTransaction>> {
-    const { token } = originTokenAmount;
-    const transactions: Array<WarpTypedTransaction> = [];
-
-    const providerType = TOKEN_STANDARD_TO_PROVIDER_TYPE[token.standard];
-
-    const transferParams = this.resolveQuotedCallsParams({
-      originTokenAmount,
-      destination,
-      recipient,
-      quotedCalls,
-      destinationToken,
+    assert(
+      isAddress(args.recipient),
+      `QuotedCalls recipient must be a valid EVM address: ${args.recipient}`,
+    );
+    const provider = new EvmQuotedTransferProvider(args.quotedCalls);
+    return provider.getQuotedTransferFee({
+      warpCore: this,
+      originTokenAmount: args.originTokenAmount,
+      destination: args.destination,
+      sender: args.sender,
+      recipient: args.recipient,
+      destinationToken: args.destinationToken,
     });
-
-    // Get fee quotes if not provided
-    if (!feeQuotes) {
-      const fees = await this.getQuotedTransferFee({
-        originTokenAmount,
-        destination,
-        sender,
-        recipient,
-        quotedCalls,
-        destinationToken,
-      });
-      feeQuotes = fees.feeQuotes;
-    }
-
-    const { tokenTotals } = extractQuoteTotals(feeQuotes);
-    const totalTokenNeeded =
-      tokenTotals.get(transferParams.token.toLowerCase() as `0x${string}`) ??
-      0n;
-
-    // Check approval for QuotedCalls (TransferFrom mode).
-    // The spender is quotedCalls.address (not the token itself), so
-    // EvmHypSyntheticAdapter correctly falls through to the ERC20 allowance
-    // check rather than returning false.
-    if (
-      quotedCalls.tokenPullMode === TokenPullMode.TransferFrom &&
-      totalTokenNeeded > 0n
-    ) {
-      const adapter = token.getAdapter(this.multiProvider);
-      const [isApproveRequired, isRevokeApprovalRequired] = await Promise.all([
-        adapter.isApproveRequired(
-          sender,
-          quotedCalls.address,
-          totalTokenNeeded,
-        ),
-        adapter.isRevokeApprovalRequired(sender, quotedCalls.address),
-      ]);
-      // USDT-like tokens require revoking to 0 before re-approving
-      if (isApproveRequired && isRevokeApprovalRequired) {
-        const revokeTxReq = await adapter.populateApproveTx({
-          weiAmountOrId: 0,
-          recipient: quotedCalls.address,
-        });
-        transactions.push({
-          category: WarpTxCategory.Revoke,
-          type: providerType,
-          transaction: revokeTxReq,
-        } as WarpTypedTransaction); // CAST: providerType is determined at runtime from token.standard
-      }
-      if (isApproveRequired) {
-        const approveTxReq = await adapter.populateApproveTx({
-          weiAmountOrId: totalTokenNeeded,
-          recipient: quotedCalls.address,
-        });
-        transactions.push({
-          category: WarpTxCategory.Approval,
-          type: providerType,
-          transaction: approveTxReq,
-        } as WarpTypedTransaction); // CAST: providerType is determined at runtime from token.standard
-      }
-    }
-
-    // Build execute tx with exact fee amounts
-    const executeTx = buildExecuteCalldata({
-      ...transferParams,
-      feeQuotes,
-      tokenPullMode: quotedCalls.tokenPullMode,
-      permit2Data: quotedCalls.permit2Data,
-    });
-
-    transactions.push({
-      category: WarpTxCategory.Transfer,
-      type: providerType,
-      transaction: {
-        to: executeTx.to,
-        data: executeTx.data,
-        value: executeTx.value.toString(),
-      },
-    } as WarpTypedTransaction); // CAST: providerType is determined at runtime from token.standard
-
-    return transactions;
   }
 
   /**
@@ -1862,42 +1642,12 @@ export class WarpCore {
     destination: ChainNameOrId;
     destinationToken?: IToken;
   }): IToken {
-    const destinationName = this.multiProvider.getChainName(destination);
-    const destinationCandidates = originToken
-      .getConnections()
-      .filter((connection) => connection.token.chainName === destinationName)
-      .map((connection) => connection.token);
-
-    assert(
-      destinationCandidates.length > 0,
-      `No connection found for ${destinationName}`,
-    );
-
-    if (destinationToken) {
-      assert(
-        destinationToken.chainName === destinationName,
-        `Destination token chain mismatch for ${destinationName}`,
-      );
-      const matchedToken = destinationCandidates.find(
-        (candidate) =>
-          candidate.equals(destinationToken) ||
-          tokenIdentifiersEqual(
-            candidate.addressOrDenom,
-            destinationToken.addressOrDenom,
-          ),
-      );
-      assert(
-        matchedToken,
-        `Destination token ${destinationToken.addressOrDenom} is not connected from ${originToken.chainName} to ${destinationName}`,
-      );
-      return matchedToken;
-    }
-
-    assert(
-      destinationCandidates.length === 1,
-      `Ambiguous route to ${destinationName}; specify destination token`,
-    );
-    return destinationCandidates[0];
+    return resolveDestinationToken({
+      multiProvider: this.multiProvider,
+      originToken,
+      destination,
+      destinationToken,
+    });
   }
 
   /**
